Adding HasteMatcher to allow for stateful matches later; allowing lookaheads in matching

+++ Assets/Haste/Editor/HasteResult.cs
@@ -6,15 +6,15 @@ namespace Haste {
   public struct HasteResult {
     public string Path;
     public HasteSource Source;
-    public int Score;
+    public float Score;
 
-    public HasteResult(string path, HasteSource source, int score = 0) {
+    public HasteResult(string path, HasteSource source, float score = 0) {
       Path = path;
       Source = source;
       Score = score;
     }
 
-    public HasteResult(HasteItem item, int score = 0) {
+    public HasteResult(HasteItem item, float score = 0) {
       Path = item.Path;
       Source = item.Source;
       Score = score;
